Now using shared pointers for snapshots. Got multithreaded change detection working (30s->10s compute? idk i haven't benchmarked).

+++ include/region_properties.hpp
@@ -26,7 +26,7 @@ struct MemoryRegionProperties {
     // Should this be const? yes.
     const std::string     parentRegionName;
     const uintptr_t       parentRegionStart;
-    const size_t          parentRegionSize;
+    const uint64_t        parentRegionSize;
     const PermissionsMask permissions;
 
   public:
